+++ src/Compute/Compute/ChangeLog.md
@@ -21,6 +21,8 @@
 -->
 ## Upcoming Release
 * Edited the documentation for the cmdlet `Set-AzVMADDomainExtension` to ensure the example is accurate. 
+* Updated New-AzVM feature for `vCPUsAvailable` and `vCPUsPerCore` parameters. Cmdlets will not try to use the new `VMCustomizationPreview` feature if the user does not have access to that feature. [#17370]
+* Improved description and examples for disk creation
 
 ## Version 4.24.0
 * Upgraded Compute .NET SDK package reference to version 52.0.0